+++ 404.php
@@ -2,7 +2,7 @@
 /**
  * The template for displaying 404 pages (not found).
  *
- * @package Theme 12
+ * @package Fremont Cut
  */
 
 get_header(); ?>
@@ -12,11 +12,11 @@ get_header(); ?>
 
 			<section class="error-404 not-found">
 				<header id="f04header" class="page-header">
-					<h1 class="page-title"><?php _e( '404 Not Found.', 'theme-12' ); ?></h1>
+					<h1 class="page-title"><?php _e( '404 Not Found.', 'fremont-cut' ); ?></h1>
 				</header><!-- .page-header -->
 
 				<div class="page-content">
-					<p><?php _e( 'Nothing was found at this location.', 'theme-12' ); ?></p>
+					<p><?php _e( 'Nothing was found at this location.', 'fremont-cut' ); ?></p>
 
 					<?php //	get_search_form(); ?>
 
@@ -24,7 +24,7 @@ get_header(); ?>
 
 					<?php if ( greenlake_categorized_blog() ) : // Only show the widget if site has multiple categories. ?>
 					<div class="widget widget_categories">
-						<h2 class="widget-title"><?php // _e( 'Most Used Categories', 'theme-12' ); ?></h2>
+						<h2 class="widget-title"><?php // _e( 'Most Used Categories', 'fremont-cut' ); ?></h2>
 						<ul>
 						<?php
 					//		wp_list_categories( array(
@@ -41,7 +41,7 @@ get_header(); ?>
 
 					<?php //  FORGET THIS
 						/* translators: %1$s: smiley */
-				//		$archive_content = '<p>' . sprintf( __( 'Try looking in the monthly archives. %1$s', 'theme-12' ), convert_smilies( ':)' ) ) . '</p>';
+				//		$archive_content = '<p>' . sprintf( __( 'Try looking in the monthly archives. %1$s', 'fremont-cut' ), convert_smilies( ':)' ) ) . '</p>';
 				//		the_widget( 'WP_Widget_Archives', 'dropdown=1', "after_title=</h2>$archive_content" );
 					?>