+++ content-none.php
@@ -4,28 +4,28 @@
  *
  * Learn more: http://codex.wordpress.org/Template_Hierarchy
  *
- * @package Theme 12
+ * @package Fremont Cut
  */
 ?>
 
 <section class="no-results not-found">
 	<header class="page-header">
-		<h1 class="page-title"><?php _e( 'Nothing Found', 'theme-12' ); ?></h1>
+		<h1 class="page-title"><?php _e( 'Nothing Found', 'fremont-cut' ); ?></h1>
 	</header><!-- .page-header -->
 
 	<div class="page-content">
 		<?php if ( is_home() && current_user_can( 'publish_posts' ) ) : ?>
 
-			<p><?php printf( __( 'Ready to publish your first post? <a href="%1$s">Get started here</a>.', 'theme-12' ), esc_url( admin_url( 'post-new.php' ) ) ); ?></p>
+			<p><?php printf( __( 'Ready to publish your first post? <a href="%1$s">Get started here</a>.', 'fremont-cut' ), esc_url( admin_url( 'post-new.php' ) ) ); ?></p>
 
 		<?php elseif ( is_search() ) : ?>
 
-			<p><?php _e( 'Nothing matched your search terms. Please try again with some different keywords.', 'theme-12' ); ?></p>
+			<p><?php _e( 'Nothing matched your search terms. Please try again with some different keywords.', 'fremont-cut' ); ?></p>
 			<?php get_search_form(); ?>
 
 		<?php else : ?>
 
-			<p><?php _e( 'It seems we can&rsquo;t find what you&rsquo;re looking for. Perhaps searching can help.', 'theme-12' ); ?></p>
+			<p><?php _e( 'It seems we can&rsquo;t find what you&rsquo;re looking for. Perhaps searching can help.', 'fremont-cut' ); ?></p>
 			<?php get_search_form(); ?>
 
 		<?php endif; ?>