+++ content-page.php
@@ -2,7 +2,7 @@
 /**
  * The template used for displaying page content in page.php
  *
- * @package Theme 12
+ * @package Fremont Cut
  */
 ?>
 
@@ -15,13 +15,13 @@
 		<?php the_content(); ?>
 		<?php
 			wp_link_pages( array(
-				'before' => '<div class="page-links">' . __( 'Pages:', 'theme-12' ),
+				'before' => '<div class="page-links">' . __( 'Pages:', 'fremont-cut' ),
 				'after'  => '</div>',
 			) );
 		?>
 	</div><!-- .entry-content -->
 
 	<footer class="entry-footer">
-		<?php edit_post_link( __( 'Edit', 'theme-12' ), '<span class="edit-link">', '</span>' ); ?>
+		<?php edit_post_link( __( 'Edit', 'fremont-cut' ), '<span class="edit-link">', '</span>' ); ?>
 	</footer><!-- .entry-footer -->
 </article><!-- #post-## -->