+++ inc/extras.php
@@ -4,7 +4,7 @@
  *
  * Eventually, some of the functionality here could be replaced by core features
  *
- * @package Green Lake
+ * @package Fremont Cut
  */
 
 /**
@@ -49,7 +49,7 @@ if ( version_compare( $GLOBALS['wp_version'], '4.1', '<' ) ) :
 
 		// Add a page number if necessary:
 		if ( ( $paged >= 2 || $page >= 2 ) && ! is_404() ) {
-			$title .= " $sep " . sprintf( __( 'Page %s', 'theme-12' ), max( $paged, $page ) );
+			$title .= " $sep " . sprintf( __( 'Page %s', 'fremont-cut' ), max( $paged, $page ) );
 		}
 
 		return $title;